64bit proofs the Build_Frame function which decodes sprite data.

+++ common/keyframe.h
@@ -19,7 +19,7 @@ extern char* TheaterShapeBufferStart;
 extern bool UseOldShapeDraw;
 }
 
-unsigned long Build_Frame(void const* dataptr, unsigned short framenumber, void* buffptr);
+uintptr_t Build_Frame(void const* dataptr, unsigned short framenumber, void* buffptr);
 unsigned short Get_Build_Frame_Count(void const* dataptr);
 unsigned short Get_Build_Frame_X(void const* dataptr);
 unsigned short Get_Build_Frame_Y(void const* dataptr);
